Release 1.2
Considering that similar issues have not occurred with the files on our testing platform, we are now uploading new script files that have been proven to run stably on our testing platform, including .csv format files of the EBoMD data set. All file addresses in the script file need to be changed to their corresponding locations on your platform.